geekbrains-bot-docker/requirements.txt
nest_asyncio==1.6.0
python-dotenv==1.0.1
python-telegram-bot==21.1.1
Requests==2.31.0
